post_service/requirements.txt
fastapi[standard]
requests
sqlmodel
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-fastapi